pyproject.toml
[project]
name = "fraud-detection-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.0",
    "joblib>=1.5.3",
    "networkx>=3.6.1",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "email-validator>=2.0.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.24",
    "bcrypt==4.0.1",
    "streamlit>=1.40.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
